requirements.txt
paho-mqtt~=1.6.1
urllib3==1.26.18
